+++ myems-web/src/components/MyEMS/CombinedEquipment/CombinedEquipmentEnergyItem.js
@@ -149,6 +149,7 @@ const CombinedEquipmentEnergyItem = ({ setRedirect, setRedirectUrl, t }) => {
 
   const [associatedEquipmentTableData, setAssociatedEquipmentTableData] = useState([]);
   const [associatedEquipmentTableColumns, setAssociatedEquipmentTableColumns] = useState([
+    { dataField: 'id', text: t('ID'), sort: true },
     { dataField: 'name', text: t('Associated Equipment'), sort: true }
   ]);
 
@@ -773,7 +774,9 @@ const CombinedEquipmentEnergyItem = ({ setRedirect, setRedirectUrl, t }) => {
             json['associated_equipment']['associated_equipment_names_array'][0].forEach(
               (currentEquipmentName, equipmentIndex) => {
                 let associated_equipment_value = {};
-                associated_equipment_value['id'] = equipmentIndex;
+                const ids = json['associated_equipment']?.associated_equipment_ids || [];
+                const rawId = ids[equipmentIndex] ?? equipmentIndex;
+                associated_equipment_value['id'] = Number(rawId);
                 associated_equipment_value['name'] = currentEquipmentName;
                 json['associated_equipment']['energy_item_names'].forEach((currentValue, energyItemIndex) => {
                   associated_equipment_value['a' + energyItemIndex] =
@@ -787,6 +790,11 @@ const CombinedEquipmentEnergyItem = ({ setRedirect, setRedirectUrl, t }) => {
           setAssociatedEquipmentTableData(associated_equipment_value_list);
 
           let associated_equipment_column_list = [];
+          associated_equipment_column_list.push({
+            dataField: 'id',
+            text: t('ID'),
+            sort: true
+          });
           associated_equipment_column_list.push({
             dataField: 'name',
             text: t('Associated Equipment'),