+++ myems-web/src/components/MyEMS/CombinedEquipment/CombinedEquipmentLoad.js
@@ -146,6 +146,11 @@ const CombinedEquipmentLoad = ({ setRedirect, setRedirectUrl, t }) => {
   const [associatedEquipmentTableData, setAssociatedEquipmentTableData] = useState([]);
   const [associatedEquipmentTableColumns, setAssociatedEquipmentTableColumns] = useState([
     {
+      dataField: 'id',
+      text: t('ID'),
+      sort: true
+    },
+    {
       dataField: 'name',
       text: t('Associated Equipment'),
       sort: true
@@ -771,7 +776,9 @@ const CombinedEquipmentLoad = ({ setRedirect, setRedirectUrl, t }) => {
             json['associated_equipment']['associated_equipment_names_array'][0].forEach(
               (currentEquipmentName, equipmentIndex) => {
                 let associated_equipment_value = {};
-                associated_equipment_value['id'] = equipmentIndex;
+                const ids = json['associated_equipment']?.associated_equipment_ids || [];
+                const rawId = ids[equipmentIndex] ?? equipmentIndex;
+                associated_equipment_value['id'] = Number(rawId);
                 associated_equipment_value['name'] = currentEquipmentName;
                 json['associated_equipment']['energy_category_names'].forEach((currentValue, energyCategoryIndex) => {
                   if (json['associated_equipment']['sub_averages_array'][energyCategoryIndex][equipmentIndex] != null) {
@@ -795,6 +802,11 @@ const CombinedEquipmentLoad = ({ setRedirect, setRedirectUrl, t }) => {
           setAssociatedEquipmentTableData(associated_equipment_value_list);
 
           let associated_equipment_column_list = [];
+          associated_equipment_column_list.push({
+            dataField: 'id',
+            text: t('ID'),
+            sort: true
+          });
           associated_equipment_column_list.push({
             dataField: 'name',
             text: t('Associated Equipment'),